sync_saihu_erp/data_update/requirements.txt
# 赛狐ERP数据同步系统依赖包

# 核心依赖
PyMySQL>=1.0.2
requests>=2.28.0
PyYAML>=6.0
python-dateutil>=2.8.2

# 任务调度
APScheduler>=3.10.1

# 数据处理
pandas>=1.5.0
numpy>=1.24.0

# 日志
colorlog>=6.7.0

# 配置管理
python-decouple>=3.6

# 时区处理
pytz>=2023.3

# 数据验证
cerberus>=1.3.4

# HTTP客户端增强
urllib3>=1.26.0
chardet>=5.0.0

# 加密支持
cryptography>=3.4.8

# 开发和测试
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=5.0.0
black>=22.0.0

# 可选依赖：监控和告警
# prometheus-client>=0.15.0
# requests-oauthlib>=1.3.0

# 可选依赖：数据库迁移
# alembic>=1.8.0